+++ includes/init_includes/init_category_path.php
@@ -35,7 +35,7 @@ if (zen_not_null($cPath)) {
   $cPath = implode('_', $cPath_array);
   $current_category_id = $cPath_array[(sizeof($cPath_array)-1)];
 } else {
-  $current_category_id = TOPMOST_CATEGORY_PARENT_ID;
+  $current_category_id = (int)TOPMOST_CATEGORY_PARENT_ID;
   $cPath_array = array();
 }
