+++ includes/templates/template_default/templates/tpl_advanced_search_default.php
@@ -31,7 +31,7 @@
 
 <fieldset class="floatingBox back">
     <legend><?php echo ENTRY_CATEGORIES; ?></legend>
-    <div class="floatLeft"><?php echo zen_draw_pull_down_menu('categories_id', zen_get_categories(array(array('id' => '', 'text' => TEXT_ALL_CATEGORIES)), TOPMOST_CATEGORY_PARENT_ID ,'', '1'), $sData['categories_id']); ?></div>
+    <div class="floatLeft"><?php echo zen_draw_pull_down_menu('categories_id', zen_get_categories(array(array('id' => '', 'text' => TEXT_ALL_CATEGORIES)), (int)TOPMOST_CATEGORY_PARENT_ID ,'', '1'), $sData['categories_id']); ?></div>
 <?php echo zen_draw_checkbox_field('inc_subcat', '1', $sData['inc_subcat'], 'id="inc-subcat"'); ?><label class="checkboxLabel" for="inc-subcat"><?php echo ENTRY_INCLUDE_SUBCATEGORIES; ?></label>
 <br class="clearBoth" />
 </fieldset>
